requirements.txt
anthropic>=0.51.0
chromadb>=1.0.0
datasets>=2.20.0
json-repair>=0.35.0
langchain-text-splitters>=0.3.0
openai>=1.75.0
python-dotenv>=1.0.1
pydantic>=2.8.0
sentence-transformers>=3.0.0
tiktoken>=0.9.0
tqdm>=4.66.0
